Ensure that MMS reception ACK is sent to prevent duplicate MMS
This implements the getMmscInfoForReceptionAck method
in MmsReceiver which is according to android-smsmms
library required for some carriers, since otherwise
MMS would be duplicated. Unfortunately,
accessing MMSC information is more restricted with
each version of Android and this implementation relies
on system database on older versions and uses APN database
included with android-smsmms for newer versions, which will
work only on the default SIM card.